# Conflicts:
#	Wisdom_Education_Server/restful_apis.md

+++ Wisdom_Education_Android/recordplay-model/src/main/java/com/netease/yunxin/app/wisdom/record/model/NERecordItem.kt
@@ -18,7 +18,7 @@ data class NERecordItem(
     val roomCid: String,
     val roomUid: Long,
     val roomUuid: String,
-    val size: Int,
+    val size: Long,
     val timestamp: Long,
     val type: String,
     val url: String,
